requirements.txt
altair==5.5.0
anyio==4.8.0
APScheduler==3.11.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bokeh==3.6.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.55.3
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.20.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.2
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tradingview-ta==3.3.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
watchdog==6.0.0
webencodings==0.5.1
xyzservices==2025.1.0
yfinance==0.2.65
